Add more session data to cookies
This adds cookies for more of the data that is sent back from the new
session request so we can use it to optionally display links and prefill
forms.

Two other approaches I can think of are storing this in local storage,
and storing this is a 'global state' with redux.

Local storage can be cleared by the browser and may not be available to
some browsers. Introducing redux could cause more confusion than it's
worth.

I'm open to suggestions for any other options.