Set auth header for GET /mentors and GET /services
Now that we've locked down the mentors and services requests, this will
add a header to those get requests to retrieve them for the mentor
request form.

This will also redirect to the login page from the mentor request page
if the mentor request page is rendered and the user does not have an
auth token.