Add autoBlur & autosize attrs to Select
autosize lets the Select grow in width when multiple values are selected. changed CSS width -> min-width to allow that.

autoBlur - from react-select docs: 	'Blurs the input element after a selection has been made. Handy for lowering the keyboard on mobile devices'

autoblur also removes focus from the select when a selection is made, so that it doesn't obscure the school cards directly below the select, if only one selection made